add files unit tests

+++ files/files.go
@@ -8,12 +8,17 @@ import (
 )
 
 func Run(c *context.Context, f config.File) error {
-	fmt.Printf("Writing to %s ... \n", f.Path)
-	data := []byte(f.Content)
-	err := ioutil.WriteFile(f.Path, data, f.Mode)
-	if err != nil {
-		return err
+	if c.IsDryRun {
+		fmt.Printf("DRY-RUN: Writing to %s ... \n", f.Path)
+		fmt.Printf("DRY-RUN: Done.\n")
+	} else {
+		fmt.Printf("Writing to %s ... \n", f.Path)
+		data := []byte(f.Content)
+		err := ioutil.WriteFile(f.Path, data, f.Mode)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Done.\n")
 	}
-	fmt.Printf("Done.\n")
 	return nil
 }
